move function to output_helpers
